Maven-checkstyle-plugin, more checkes

+++ src/test/java/com/epam/tc/hw1/AddTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class AddTest extends AbstractTest {
     @Test(dataProvider = "sumData", dataProviderClass = DataProviders.class)
-        public void calcTest(int a, int b, int expected) {
+        public void calcTest(long a, long b, long expected) {
         long result = calculator.sum(a, b);
         assertThat(result).as("Addition calculated incorrectly").isEqualTo(expected);
     }
