+++ src/test/java/com/epam/tc/hw1/DivTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class DivTest extends AbstractTest {
     @Test(dataProvider = "divData", dataProviderClass = DataProviders.class)
-    public void calcTest(int a, int b, int expected) {
+    public void calcTest(long a, long b, long expected) {
         long result = calculator.div(a, b);
         assertThat(result).as("Divide calculated incorrectly").isEqualTo(expected);
     }
