+++ src/test/java/com/epam/tc/hw1/MultTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class MultTest extends AbstractTest {
     @Test(dataProvider = "multData", dataProviderClass = DataProviders.class)
-    public void calcTest(int a, int b, int expected) {
+    public void calcTest(long a, long b, long expected) {
         long result = calculator.mult(a, b);
         assertThat(result).as("Multiplication calculated incorrectly").isEqualTo(expected);
     }
