+++ src/test/java/com/epam/tc/hw1/SubsTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class SubsTest extends AbstractTest {
     @Test(dataProvider = "subsData", dataProviderClass = DataProviders.class)
-    public void calcTest(int a, int b, int expected) {
+    public void calcTest(long a, long b, long expected) {
         long result = calculator.sub(a, b);
         assertThat(result).as("Subtraction calculated incorrectly").isEqualTo(expected);
     }
